requirements.txt
pytest==8.4.1
aiogram==3.22.0
python-dotenv==1.1.0
psycopg2==2.9.10
python-telegram-bot==22.3
pytest-cov==6.2.1
langchain-core==0.3.75
langchain-gigachat==0.3.12
mistralai==1.9.10